ats-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
django-filter==23.3
Pillow==10.1.0
psycopg2-binary==2.9.8
celery==5.3.4
redis==5.0.1
boto3==1.34.0
PyPDF2==3.0.1
python-docx==1.1.0
